+++ tests/functional/test_http_validate.py
@@ -1,52 +1,36 @@
-from unittest.mock import patch
-
 import pytest
+from click.testing import CliRunner
 
 from pyrandall import cli
-from tests.conftest import vcr
 
-CONFIG = "examples/config/v1.json"
 ARGV_HTTP_VALIDATE_1_OK = [
     "--config",
-    CONFIG,
-    "--dataflow",
-    "examples/",
-    "sanitytest",
-    "http/validate_ok_status_code.yaml",
+    "examples/config/v1.json",
+    "-V",
+    "examples/scenarios/http/validate_ok_status_code.yaml",
 ]
 ARGV_HTTP_VALIDATE_STAUTS_CODE_FAIL = [
     "--config",
-    CONFIG,
-    "--dataflow",
-    "examples/",
-    "sanitytest",
-    "http/validate_bad_status_code.yaml",
+    "examples/config/v1.json",
+    "-V",
+    "examples/scenarios/http/validate_bad_status_code.yaml",
 ]
 
 
-@patch("pyrandall.cli.ArgumentParser._print_message")
-def test_execute_a_sanitytest_fails(print_message):
-    with pytest.raises(SystemExit) as context:
-        cli.start([])
-    assert context.value.code == 2
-    print_message.assert_called()
+def test_validate_assertions_pass(vcr):
+    with vcr.use_cassette("test_validate_assertions_pass") as cassette:
+        runner = CliRunner()
+        result = runner.invoke(cli.main, ARGV_HTTP_VALIDATE_1_OK, catch_exceptions=False)
+        assert 'Usage: pyrandall' not in result.output
+        assert result.exit_code == 0
+        assert cassette.all_played
 
 
-@vcr.use_cassette("test_validate_assertions_pass")
-def test_validate_assertions_pass():
-    with pytest.raises(SystemExit) as context:
-        cli.start(ARGV_HTTP_VALIDATE_1_OK)
-    if context.value.code == 2:
-        pytest.fail(cli.argparse_error(ARGV_HTTP_VALIDATE_1_OK))
+def test_validate_fail_status_code(vcr):
+    with vcr.use_cassette("test_validate_fail_status_code") as cassette:
+        runner = CliRunner()
+        result = runner.invoke(cli.main, ARGV_HTTP_VALIDATE_STAUTS_CODE_FAIL, catch_exceptions=False)
 
-    assert context.value.code == 0
-
-
-@vcr.use_cassette("test_validate_fail_status_code")
-def test_validate_fail_status_code():
-    with pytest.raises(SystemExit) as context:
-        cli.start(ARGV_HTTP_VALIDATE_STAUTS_CODE_FAIL)
-    if context.value.code == 2:
-        pytest.fail(cli.argparse_error(ARGV_HTTP_VALIDATE_STAUTS_CODE_FAIL))
-
-    assert context.value.code == 1
+        assert 'Usage: pyrandall' not in result.output
+        assert result.exit_code == 1
+        assert cassette.all_played