+++ tests/functional/test_kafka_simulate.py
@@ -2,44 +2,29 @@ import os
 import pytest
 from freezegun import freeze_time
 
-from pyrandall import cli
-from tests.conftest import vcr
 from tests.helper import KafkaConsumer
 
 
 TEST_TOPIC = "pyrandall-tests-e2e"
+ARGV_SMALL = [
+    "--config",
+    "examples/config/v1.json",
+    "-s",
+    "examples/scenarios/v2_ingest_kafka_small.yaml"
+]
 
 
-config = "examples/config/v1.json"
-MOCK_ARGV = ["--config", config, "--dataflow", "examples/", "simulate"]
-
-
-def test_invalid_kafka_scenario():
-    argv = MOCK_ARGV + ["v2_ingest_kafka_invalid.yaml"]
-    with pytest.raises(SystemExit) as context:
-        cli.start(argv)
-    if context.value.code == 2:
-        pytest.fail(cli.argparse_error(argv))
-    assert context.value.code == 4
-
-
-# freeze time in order to hardcode timestamps
 @freeze_time("2012-01-14 14:33:12")
-@vcr.use_cassette("test_ingest_to_kafka")
-def test_simulate_produces_event(kafka_cluster_info):
+def test_simulate_produces_event(kafka_cluster_info, pyrandall_cli):
     consumer = KafkaConsumer(TEST_TOPIC)
     try:
         highwater = consumer.get_high_watermark()
 
         # run simulate to create a message to kafka
         # running following command:
-        argv = MOCK_ARGV + ["v2_ingest_kafka_small.yaml"]
-        print(f"running {argv}")
-        with pytest.raises(SystemExit) as context:
-            cli.start(argv)
-        if context.value.code == 2:
-            pytest.fail(cli.argparse_error(argv))
-        assert context.value.code == 0
+        result = pyrandall_cli.invoke(ARGV_SMALL)
+        assert 'Usage: main' not in result.output
+        assert result.exit_code == 0
 
         messages = consumer.get_messages(expecting=2)
         assert len(messages) == 2